introduce convention plugins for better build logic structuring
