+++ build-logic/src/main/kotlin/keycloak-mock.base.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    idea
+}
+
+repositories {
+    mavenCentral()
+    google()
+}
+
+group = "com.tngtech.keycloakmock"
+version = rootProject.version
