+++ build-logic/src/main/kotlin/keycloak-mock.java-library.gradle.kts
@@ -1,0 +1,5 @@
+plugins {
+    `java-library`
+    id("keycloak-mock.base")
+    id("keycloak-mock.testing")
+}
